Changed path and filenames from vehicles to vehicle, VehiclesService to VehicleService; Added tsdoc and retrieve method;
