+++ web/src/app/shared/services/vehicle/vehicle.model.ts
@@ -1,0 +1,45 @@
+/**
+ * A base class for the vehicle entity.
+ *
+ * @category Model
+ */
+export class Vehicle {
+    /**
+     * The constructor of the `Vehicle` class.
+     *
+     * @param id the vehicle's database ID
+     * @param plateNo the vehicle's plate no.
+     * @param vehicleType the vehicle's type {@link VehicleType}
+     * @param resident the vehicle's resident ID 
+     * @param isActive the vehicle is active or not? 
+     * @param createdAt the vehicle's entry creation date and time 
+     * @param createdBy the vehicle's entry creator ID
+     * @param lastModifiedAt the vehicle's entry last modification date and time
+     * @param lastModifiedBy the vehicle's entry last modificator ID
+     */
+    constructor(
+        public id: number,
+        public plateNo: string,
+		public vehicleType: VehicleType,
+		public resident: boolean,
+		public isActive: boolean,
+        public createdAt: string,
+        public createdBy: number,
+        public lastModifiedAt: string,
+        public lastModifiedBy: number
+    ) {}
+}
+
+/**
+ * An enum for vehicle types
+ */
+export enum VehicleType {
+    /** A car type vehicle */
+    CAR = 1,
+
+    /** A motorcycle type vehicle */
+    MOTOR = 2,
+
+    /** A lorry type vehicle */
+    LORRY = 3
+}
